requirements.txt
praw>=7.0.0
yfinance>=0.1.63
vaderSentiment>=3.3.2
pandas>=1.2.0
nltk>=3.6.0
python-dotenv>=0.19.0